Made notebooks run again.
